requirements/base.txt
aiohttp==2.2.0
aiohttp-debugtoolbar==0.4.0
aiohttp-devtools==0.5
aiohttp-jinja2==0.13.0
aiohttp-transmute==1.1.0
aiopg==0.13.0
alembic==0.8.6
argh==0.26.2
async-timeout==1.2.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
decorator==4.0.11
httpio==0.1.3
idna==2.5
isort==4.2.15
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
migrate==0.3.8
multidict==3.1.0
pathtools==0.1.2
pbr==3.1.1
psycopg2==2.6.1
python-dateutil==2.6.0
python-editor==1.0.3
PyYAML==3.12
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
urllib3==1.21.1
watchdog==0.8.3
yarl==0.10.3
